Fix compile warnings

+++ XboxInternals/Stfs/StfsDefinitions.cpp
@@ -215,7 +215,7 @@ void WriteCertificateEx(Certificate *cert, BaseIO *io, DWORD address)
     io->Write((WORD)cert->publicKeyCertificateSize);
     io->Write(cert->ownerConsoleID, 5);
     io->Write(cert->ownerConsolePartNumber, 0x11, false);
-    DWORD temp = cert->consoleTypeFlags | cert->ownerConsoleType;
+    DWORD temp = static_cast<DWORD>(cert->consoleTypeFlags) | static_cast<DWORD>(cert->ownerConsoleType);
     io->Write(temp);
     io->Write(cert->dateGeneration, 0x8, false);
     io->Write(cert->publicExponent);
